Remove hard-coded /usr/local/share plugin path
The build added /usr/local/share/usd/plugins as a
hard-coded plugin search path, the idea being that
consumers could add their own site-wide USD plugins
to that area. However, this hard-coded path made it
impossible to create a completely self-contained
build, even when the build was relocated to a new
location.

So, this hard-coded path is being removed. Consumers
that want to establish a site-wide plugin location
can add that to the plugin search path via the
PXR_INSTALL_LOCATION cmake option.

(Internal change: 1698024)