Fix for incorrect lifetime management of texture resources.
TestUSDInvalidTexPath was failing due to GL errors.  The cause of these errors was that a surface shader was binding an deleted sampler.  This regression was caused by introducing garbage collection on texture resources.
The was the resource instance registries are designed to work, is using the shared_ptr class to do implict ref counting.  When the instance registry holds the only remaining reference, the object is removed from the cache.  As such the Prims that use the instances are expected to hold onto a shared_ptr while they are using the instance.
The problem in this case is HdTexture was never holding onto the shared_ptr to the HdTextureResource.  So the reference count was always 1, so the garbage collector deleted it.

Changes:
- If HdTexture creates the texture resource, it now uses the class member instead of a local varable.
- If HdTexture is updating and hits the cache, it update the class member to the hit resource (effectively referencing the current resource and increasing the reference on the hit object).

(Internal change: 1698095)