Add Bprims to Hydra.
This change is part of the Hydra refactor.

A Bprim (buffer prim) is a class of prim for managing a blob of data that is
used to communicate between the scene delegate and render.

Like other prim types (Rprim and Sprim), the Bprim communicates with the
scene delegate got get buffer properties (e.g. the size of the buffer) as
well as the contents of the buffer.

Changes to the properties and contents are change tracked and updates
are cached in the renderer.  The Render Delegate may choose to transform
the data into a renderer specific form on download.

BPrims are sync'ed first and thus, Bprims should not be
Dependent on the state of any other prim.

The most typical use of a Bprim would be a Texture and it is planned to later change HdTexture to use the Bprim API.

Changes:
- Add Bprim base class
- Added Bprim support to change tracjer
- Added Bprim support to render index.
- Added Bprim allocation support to Render delegate.

Notes:
Bprim sync has not been implmeneted in this change and will come later.

(Internal change: 1698189)