Convert HdTexture to a Bprim
This change is part of the Hydra refracting effort.
Texture's are a type of Bprim (buffer primitive), this change makes HdTexture derive from Bprim.  It also removes specific texture tracking for the change tracker and render index, instead utilizing the new bprim tracking.
The changes to the Sync are transitory as a bigger cleanup of sync is planned.

Changes:
- HdTexture now derives from Bprim.
- Added doxygen comments for HdTexture
- Removed Texture specific api from change tracker and render index
- HdTexture class now owns it's dirty bits (matching sprim pattern).
- Updated delegates to use Bprim api to create textures.

(Internal change: 1698589)