[UsdImaging] generalize nested instancing
This change makes it possible to have heterogeneous nesting
of "instancing" adapters.
There used to be some code around in pointInstancerAdapter
which assumes the parent instancer is always pointInstancer
if it's not native instancing. This change relaxes that limitation
by delegating necessary operations of nested instancing
population/update/picking to the appropriate adapter.
Following APIs are added to primAdapter.

- GetInstanceIndices(instancer, protoInstancer)
- GetRelativeInstancerTransform(instancer, protoInstancer, time)
When a protoInstancer is instanced by another parent instancer,
protoInstancer (child) calls these API on the parent instancer to
get the instance index array and instancerTransform for itself.

- GetPathForInstanceIndex(...)
when resolving a picking result, UsdImagingDelegate recursively
calls GetPathForInstanceIndex. During the resolution of a level
of instancing, if there is one or more parent instancers to be
resolved, find adapter for the parent instancer and redirect the
process.

note: there's a bug of GetRelativeInstancerTransform that
intentionally ignores the transform of prototype root prim, which
was inherited from preceding implementation of point instancer.
It will be fix to correctly reflect the UsdGeomPointInstancer schema
in the later change.

(Internal change: 1699145)