translate lighting directly from Maya MDrawContext to GlfSimpleLightingContext
This change implements a direct translation from lighting information stored
in Maya's MDrawContext to a GlfSimpleLightingContext when using VP2.0. With
this, we no longer need to use OpenGL as an intermediary between the two
lighting systems.

For VP1.0, we still need to pass the lighting through OpenGL, as there is
no MDrawContext available.

(Internal change: 1699168)
(Internal change: 1701905)