Expand UsdImaging's Collections to allow enabling/disabling of purpose=proxy geometry independently of other purposes. Previously there was no way to disable drawing of proxy, which was a nuisance for debugging 'render' geometry in usdview (since proxy geometry frequently obscures render), and disabling proxy geometry is also a prerequisite for usdImaging being able to feed "final quality" render backends that will often just want render geom, not proxy.
Enhance usdview to be able to independently toggle proxy.  Note that
we still do not allow you to toggle "default" off.  The way that
purpose is handled, via "collections" in usdImaging, currently, would
require some very ugly, overly-verbose constructs to support yet
another dimension of variability, and I don't think it's worth it for
something we'd rarely want to toggle.

(Internal change: 1699416)