Added Windows/OS X support for ARCH_{CONSTRUCTOR,DESTRUCTOR}.
On Linux we can use the constructor(<priority>) attribute but on
OS X the priority is ignored across translation units and Windows
has no comparable attribute.

This change puts ARCH_CONSTRUCTOR() and ARCH_DESTRUCTOR() functions
on OS X and Windows into custom sections of the library and installs
code to invoke those functions in priority order when the library is
loaded or unloaded.

This change required changing how those macros are used.  Instead of
being able to simply mark a function the macros now need to declare
the function and take a pointer to it.  So the macros now take the
function name as an argument and all the clients have been suitably
updated.

Fixes #117

(Internal change: 1702304)