Enhancing UsdShade API to allow creating and connecting outputs on subgraphs and shaders.
Added a UsdShadeOutput wrapper which wraps an output on a shader or a subgraph.
Currently, an output is represented by an (attribute, relationship) pair and is connectable
via the relationship.

Retire the concept of terminals on subgraphs in favor of outputs. The associated API was
not in use outside of unit tests in usdShade. There's API in usdRi that creates terminal-like
relationships (eg, UsdRiLookAPI::{GetBxdfRel,GetDisplacementRel} etc.), but we're hopeful
that  those can be replaced with UsdShadeOutputs as well in the future. Additionally, they do
not make use of the core "terminals" API and don't create relationships in the "terminals:"
namespace.

Added an API class UsdShadeConnectableAPI which handles connecting UsdShadeParameters
and UsdShadeOutputs on subgraphs and shaders. UsdShadeSubgraph and UsdShadeShader
implicitly convert to UsdShadeConnectableAPI, so none of the existing API interface needs to
change.

(Internal change: 1702911)
(Internal change: 1702912)
(Internal change: 1703006)