Intergrate Basis Curves and Mesh Topologies to hdSt library
Continuing to seperate Hydra Core from Render Delegate functionality.  This change focuses on the mesh and basis curves topologies.

Due to some tight coupling with the mesh topology, some subdivision support also needs to be moved in this change.

Topology is currently used by the front-end Scene Delegate and has Render Delegate specific extensions.  Currently the topology returned by the scene delegate is duplicated.  This change takes advantage of this existing duplication to create the Render Delegate specific Topology.  However, as the resource registry is still in Hd, it still referes to the Hd version of Topology, so some type casting was needed.

Changes:
- Removed Computations from Hd version of Topology.
- Fixed some pass by value occurances in the topology API.
- Moved quadrangulate, triangluate, subdivsion and subdivision classes to hdSt library.
- Added implict operator = to HdMeshTopology (The compiler generated one was being used).
- Cleaned up header files used by topology, which meant some additional includes and forward declarations needed throughout.

(Internal change: 1704352)