Sdf: Fix O(n) overhead for adding a spec.
Creating 100k sibling SdfPrimSpecs took 190 sec, now 0.9 sec.
The key issue was that we represent children in the spec tree
as a vector held within a VtValue.  VtValue does copy-on-write
sharing, so to avoid O(n) work when adding a spec we must
avoid faulting the copy-on-write behavior.

(Internal change: 1704884)