Fixed ARCH_CONSTRUCTOR/ARCH_DESTRUCTOR in Windows release builds.
Optimization was discarding the structures emitted to the .pxrctor
and .pxrdtor sections and the Arch_ConstructorInit object as being
unused (despite the last being an object with non-trivial c'tor
and d'tor).

We prevent the structures from being discarded by making them
extern and avoid name collision by putting them in an anonymous
namespace.  We prevent the Arch_ConstructorInit being discarded
by using a pragma to force its inclusion.  Since the pragma takes
the symbol name we make it extern "C" to avoid name mangling.

(Internal change: 1704977)