[usdMaya] connect/disconnect assembly node time to global Maya time on Playback representation activate/inactivate
This change allows Maya's playback head time to propagate through to the
proxy node(s) and ultimately Hydra so that animation is supported with
assemblies in the Playback representation. Since there is Maya overhead in
having the assemblies and proxies driven by Maya's global time, the
connection is only made when the Playback representation is activated, and
it is disconnected when the Playback representation is inactivated.

(Internal change: 1704996)