Remove HdSceneDelegate pointer storage from Hydra Prims.
Hydra's R, S and B prims get passed a Scene Delegate pointer on construction.  The prims stored this pointer for use during Sync.

While preparing to refactor HdRprim into Core and Stream specific parts, it was discussed how the stream specific parts were going to get access to this delegate pointer.  The simplest posiblity was storing the pointer.  However, that would mean storing 16 bytes per Rprim for the delegate (which are commonly the same).  Delegate has a timing issue too, as the scene delegate's are expect to be notified before the prims can call the get functions.

Therefore, it was decided to remove the storage completely and instead pass the scene delegate as a parameter to the Rprim as part of the Sync process.

The render index is now responcible for providing an assosicate between a prim and the corrisponsing scene delegate.  It previous did store the scene delegate id per prim, but now stores the pointer itself.  S and B prim maps have been extended to follow the same pattern as Rprim and also store this information.

In order to support this change, a couple of new entry points were needed in the Render Index to support external clients that previous got the delegate from the Rprim.  First is a direct query to get the scene delegate of an Rprim at a specific path.  The second quieries the scene delegate and instancer ids of an Rprim.

Changes:
- Remove Scene Delegate as a parameter from Prim constructors and as a attribute of the Prims.
- R, S and B Prim Sync() method now take the scene delagate as a parameter.
- R Prims GetDrawItems() method also takes the scene delagate as a parameter, as it also performs a sync.
- Many functions that used the delegate now take the delegate as a paramter.
- Removed GetDelegate(), _GetRenderIndex() and _GetChangeTracker() methods from Rprim.

(Internal change: 1705260)