Pcp: Unified task queue
Github issue 156 reveals a case where we did not properly handle
variants.  The key issue is that variant sets can be decided
using non-local information which is gradually discovered as
the graph is built.  To properly handle this we split variant
eval into phases:

1) discover variant sets (vsets) available at a site
2) evaluate the authored selection for a vset
   (which may originate from another site)
3) evaluate the fallback for a vset
   (if no authored selection)
4) as new graph structure is discovered, revisit sites
   that have vsets but previously failed phases 2 & 3

We already had a fairly complex queue structure, so to do
this while also simplifying things we replace the queues in
Pcp_PrimIndexer with a unified queue.

This queue uses a PriorityOrder that embodies, in one place, all
ordering constraints in the evaluation algorithm.  Importantly,
it also lets us relax ordering constraints where the computed
results don't depend on them.  Ex: we no longer need to order
decorated arcs when decorators are not present.  This reduces
the call count to PcpCompareNodeStrength() from 2M -> 170k,
amounting to a cpu core time diff of 1.4s -> 0.1s.

(Internal change: 1706795)