Revert change to deterministic master assignment
The instance cache cannot arbitrarily pick a prim index
from the list of registered indices to assign to a
master; it must always use the first index in the list.
This is because USD will only compose descendant prim
indexes under this index. So, we need to undo this
portion of the change.

(Internal change: 1707400)