Merge Windows port into tf.
Besides converting a handful of logical 'and', 'or' and 'not' to
'&&', '||' and '!', respectively, this change adds TF_API to all
symbols that need it, uses arch functions to reduce the number of
places that depend on the platform/compiler, and makes a few other
changes:

TF_API_TEMPLATE_{CLASS,STRUCT} depend on whether we're importing or
exporting on Linux so they're now implemented in terms of new macros
ARCH_EXPORT_TEMPLATE and ARCH_IMPORT_TEMPLATE.

The TF_REFPTR_CONST_VOLATILE_GET() is used to work around an MSVC bug
(it fails to instantiate a template correctly).

TfFindLongestAccessiblePrefix() is now a private implementation detail.

TfEnvSetting doesn't use const members to work around an MSVC bug (it
makes the initialization dynamic instead of static).

Some values are explicitly cast to avoid compiler warnings.

Some names were changed to avoid conflict with Windows symbols.

Static tokens required new macros to support import/export.

TfSingleton now uses a mutex pointer and constructs the mutex on
demand.  Using a mutex directly causes a crash at startup because
on Windows std::mutex::mutex() isn't constexpr like it should be.

(Internal change: 1712574)