[usdMaya] create reference assemblies targetting assets directly for non-assembly imports
When an import is triggered on behalf of an assembly (usually due to
representation activation), we want any nested assemblies to use the same
filePath as the top-level assembly and base their (non-root) primPaths on
the paths in that stage. This is safe to do in terms of re-export, because
the export traversal will prune all nodes below the first level of
assemblies it encounters.

However, for non-assembly imports, any assembly nodes we create will be in
that first level of assemblies. As a result, we want them to target the
asset file directly using a root prim path to ensure that a re-export of
the Maya scene will author valid references, since USD references cannot
target non-root prim paths.

(Internal change: 1712627)