Optimization for UsdImagingInstanceAdapter::GetPathForInstanceIndex().
Profiling of an Internal tool, revealed that a very large portition of time was being spent in the SdfPath::GetPrimPath() function within the inner loop of UsdImagingInstanceAdapter::GetPathForInstanceIndex().  The main reason for this call was to do classification on the instancer to see if it was a leaf instancer or not, to see if the operation needed to be delegated to the child instancer.

This classification was already being done as part of inserting the instancer, so rather than use the child rprim map, a new set has been created that contains all the paths that fulfill this critera.

This has 3 advantages to the GetPathForInstanceIndex() function:

- Removes the expensive GetPrimPath() call from the inner-loop.
- Reduces the amount of data GetPathForInstanceIndex() has to search through to do the classification.
- Reduces the search algorthm complexity from O(n) to O(log n).

Runtime for this internal tool in production shots went from 7.5 hours down to 0.5 hours.

Changes:
- _InstancerData now contains a set of paths to child instancers.
- _InsertProtoRprim now returns a Boolean specifying whether the the instancer being added is a leaf instancer.
- UsdImagingInstanceAdapter::Populate() adds child instancer paths to the set.
- GetPathForInstanceIndex() now uses that set to see if delegation is required.

(Internal change: 1712628)