Use 64-bit hash functions for HdTopology hashes.
The topology cache in HdResourceRegistry had cache collisions for a test file with 30k prims (so 30k topologies). Birthday-problem math says that with a 32-bit hash, this should happen ~10% of the time.  With a 64-bit hash we have lots of headroom (we'd need 2 billion items for the same failure rate).

Changes:
- HdTopology::ID is a uint64_t now, instead of size_t (actually the same size, but explicitly sized now).
- Topology hash functions use ArchHash64 instead of ArchHash.

(Internal change: 1713106)
(Internal change: 1713335)