Timing updates.
Use bogomips to determine the TSC rate for Linux + intel.
Switched rdtsc to rdtscp - probably not much of a difference but it will prevent instruction reordering around the timestamp. Technically we should put a fence after the rdtscp but I haven't done that for now.

Use QueryPerformanceCounter to get tick for Windows (QueryPerformanceFrequency _should_ be the TSC rate but it isn't guaranteed to be. QPC should just issue rdtscp so this really isn't changing much, other than play nice with windows).

(Internal change: 1713197)