Convert Shaders into an Sprim.
This change is part of the Hydra Render Delegate refactoring.

This change modifies shader's so they are now stored in the render index as an Sprim.  In order to do this a new prim type has been added.  The schema class has been added as HdShader that derives from HdSprim.  The stream implementation of the prim is in HdStShader.

The new class HdStShader is a functionality wrapper around HdSurfaceShader (the previous class stored in the render index).  So now HdStShader implements the prim interface, such as Sync functionality.  HdSurfaceShader implements HdShaderCode interface used by the code generator.

The first gotcha, with this change was the handling of the "fallback shader".  This was being stored in the render index and had two use cases.  Firstly, it was being used for error handling when a reference to a shader was invalid.  The second was to use as a replacement shader when custom shading functionality was disabled.

In order to handle the first case, the render index now has a new concept of a fallback prim.  This applies to every S and B prim type supported by the render delegate.  The fallback prims are created by the render delegate (via a new API included in this change) when the render delgate is assigned to the render index. Hydra context creation will fail if the fallback prims can't be created.  Render index prim getter methods, continue to return nullptr and it up to the user of this methods to determine if the fallback prim should be used instread.

As a concequence of this change, Hydra Core (hd) will now give a coding error if trying to create an S or B prim whose type id is unsupported (previously the render delegates gave this error).

Shader reload functionality became a little bit more complex as a result fo this change and a tempory API has been put in place in order to maintain current functionality.

As S and B prim storage in the render index is nearly identical, the major difference being sync timing, a new template class has been added to provide the common functionality, rather than duplicate it.

Another gotcha's found, was to do with Sync timing.  Shader's depend on Texture's being synced.  As Textures are a B-prim, this is an expected and valid dependancy.  However, Sprims are currently synced before B-prims.  So some workaround have been put in place during this change until the sync order issues can be addressed.

The thrid gotcha was the the change tracker uses's Rprim's "AllDirty" mask for testing is the prim is clean.  This mask ignores the "Varying" bit (bit 1), which means the Sprim needed to leave that hole for change tracking to work.  This is also planned for fixing in a subsequent checkin.

Changes:
- Scene delegates updated to use Sprim API on render index instead of Shader API.
- Shader APIs removed from render index and change tracker.
- Added HdShader class as a schema for the new shader Sprim.
- Shader Prim's now define their own dirty bits in the schema class and no longer share dirty bits with Rprims.
- Added ability to get original shader filename from GlfGLSLFX
- Downstream of the prim now only uses the HdShaderCode interface instead of HdSurfaceShader directly.
- Moved the CanAggregate function to HdShdaerCode
- Added Sprim Sync to legacy HdEngine::Draw path needed for tests.
- Added reload support to HdGLSLFXShader
- The Render Delegate interface now has methods for retuning the list of S and B prim types it supports.
- The Render Delegate interface now has methods for creating fallback prims.
- Moved common S and B prim render index code into Hd_PrimTypeIndex.
- Added prim type initalization and fallback prim support to render index.

(Internal change: 1713763)
(Internal change: 1713805)
(Internal change: 1713896)