Consolidate dirty bot propagation to Sync Time.
This change is part of the Hydra Render Delegate refactoring.

DirtyBit propagation (I.e setting additional dirty bits based on those that are set) was currently happening in two places.  In the change tracker when a prim was marked dirty and in the prims themselves at sync time.

It is desriable to remove prim specific knowledge from the change tracker, so this change consolidates the dirty bit propagation to within the prims.

Changes:

- moved _PropagateDirtyBits from HdChangeTracker to HdRprim and renamed to _PropagateRprimDirtyBits.

- Added calls to _PropagateRprimDirtyBits from each rprim.

(Internal change: 1713771)