Fixed testPlug.
When testPlug was changed to use the python unittest module
the all important call to unittest.main() wasn't added, meaning
none of the tests were actually being run.

Furthermore, the tests were broken.  unittest runs tests in
sorted name order but that wasn't the expected behavior.  The
tests were changed to be independent using a setUpClass()
method to register the plugins and letting the individual tests
go from there.

However that still caused problems because of bugs added prior
to the migration to unittest.  Some Python plugins were declaring
a type 'PlugTypeX' in the plugInfo.json but then using
Tf.Type.Define(PlugTypeX) where PlugTypeX is a class in the
Python module.  Tf.Type.Define names such types ModuleY.PlugTypeX
leading to two type PlugTypeX and ModuleY.PlugTypeX that are not
the same and only the former being known to the Plug.Plugin.  The
plugInfo.json files, GetTypeName() implementations, and testPlug.py
were fixed to use types of the form ModuleY.PlugTypeX where
appropriate.

Still the tests were broken because the CMake rule for C++ test
shared libraries did not install plugInfo.json files so the
plugins could not be found.  The necessary code was added to
mirror both the test module install code (for filesystem locations)
and the plugin install code (for string replacement).  The
plugInfo.json files were edited to accept some string replacements
so the installed shared libraries can be located.

(Internal change: 1713799)