[Pxr] Remove BOOST_STATIC_ASSERT usage.
C++11 provides static_assert as a standard construct.
Wherever possible, we should prefer these to counterparts
from the boost libraries.

Note that the nullptr replacement in plug/staticInterface.h is
due to the fact that the removal of the boost header removes the
definition of the NULL macro. Rather than bringing in something
like <cstdlib> to bring the definition back into scope, I chose to
just replace it with its typesafe cousin from C++11.

(Internal change: 1715152)