Added UsdShadeInput which is intended to replace both UsdShadeParameter on shaders and UsdShadeInterfaceAttribute on subgraphs. These are represented attributes in the "inputs:" namespace.
Reducing the surface area of connection editing API by implementing it entirely in the API
schema UsdShadeConnectableAPI. Connections on inputs and output can only be authored
through the use of this API. This allows us to change the way connections are encoded
in shading networks without modifying client code in the future. Older usdShade objects
like UsdShadeParameter and UsdShadeInterfaceAttribute continue to have specialized
connection authoring API, but they will be deprecated very soon.

Additionally, the connection authoring API on UsdShadeConnectableAPI is fully backwards
compatible. i.e. it authors shading networks with the new encoding only if the envSetting
"USD_SHADE_WRITE_NEW_ENCODING" is enabled. Additionally, GetInput(s) will return an
interface attribute on a subgraph or a parameter on a shader (wrapped in a UsdShadeInput
object), if it doesn't find an input in the "inputs:" namespace and if the envSetting
"USD_SHADE_READ_NEW_ENCODING" is enabled. USD_SHADE_READ_NEW_ENCODING is enabled
by default. USD_SHADE_WRITE_NEW_ENCODING is disabled currently.

Added API on UsdShadeSubgraph to compute the mapping from interface inputs on a subgraph
to their  corresponding consumers. This API can optionally follow connections to inputs on
subgraphs recursively down to the shader inputs and return a transitive mapping of inputs to
consumers.

(Internal change: 1715942)