Removal of Shader/Texture Id and Mask dirty bits from HdSyncRequestVector.
This change is part of the Hydra Render Delegate refactoring.

HdSyncRequestVector structure is passed to a scene delegate's Sync API function.

It currently contains a list of surfaceShaderIDs and textureIDs that are dirty and require syncing.  Calling out these two prim types doesn't make sense with the render delegate refactor with more generic prim classes, so this API requires modification.  Furthermore, having to provide this information to the delegate and following the rule the delegate gets called first is a hinderence to current Sync function prim re-ordering nessasary to get rid of specialized code in Sprim refactoring.

As these Id's are not currently used by the Usd or any of our internal scene delegates, the Id lists are being removed.

The maskedDirtyBits member of this structure was also marked with an XXX to be removed as Hydra no-longer used it as documented.  As there is now only 1 dirty bit value in the structure it's been renamed to dirtyBits.

Because the maskedDirtyBits were expected to be updated, the USD delegate managed dirtybits by keeping a pointer to the original value.  As they are no longer needed to be updated, some of the pointer usage was removed.

Changes:
- Moved  surfaceShaderIDs and  textureIDs from HdSyncRequestVector to sync's internal structure _RprimSyncRequestVector.  They can be removed completely during sync reording.
- Removed dirty bit mask.
- Modified USD Imaging delegate to use the allDirtyBits instead of the mask.

(Internal change: 1716210)