Remove SyncSprim from testHdxCameraAndLight
This test was using Sprim's but also using the depricated HdEngine::Draw() API.  At the time Draw didn't sync Sprims.  To workaround this, the test manually synced Sprims in the render index.

When the change was made to convert Shaders to Sprims, SyncSPrims() was added to that draw path.  Therefore, this test no longer needs to sync Sprim's itself.

Changes:
- Removed calls to  SyncSPrims()

(Internal change: 1717376)