Consolidate Draw & Execute paths into a single code path.
This change is part of the Hydra Render Delegate refactoring.

While context-based entry point API(s) for this algorithm are still being discussed, with the name and parameter list not final, this consolidates the code paths internally, so we only have a single implementation of the algorithm to move forward with other sync refactoring work.

So in this change, the previously deprecated api, Draw() has been reimplemented in terms of Execute.  It those this by creating a simple task and passing that to the Execute() function.

Changes:
- Added private class Hd_DrawTask, which Draw creates on the fly and passes to Execute.

(Internal change: 1717379)