Extend Prim Sync API's.
This change is part of the Hydra Render Delegate Refactor.

A Render Delegate is responsible for Syncing Prims rather than the Core Hd library.  In order to achieve this, some changes were needed to the current Sync API's.

The first issue that needed to be addresses was that Prim's may need access to some Render Delegate shared state.  In order to achieve that, Core Hd queries the render delegate for a state object using the new GetRenderParam() API.

Sprim and Bprim sync, now take the render param as well as the dirty bits expected to be synced.  The dirty bits were added to avoid an additional change tracker lookup to fetch the dirty bits (as the core algroithm has already done this for the prim).

Rprim sync function is now pure virtual and takes the render param.  The functionality in the previous Rprim sync function has mooved into a protected function and is called from the specific prim's implementation of the Sync function.

Changes:
- Updated param list in Sprim and Bprim Sync functions.
- Hd_PrimTypeIndex's SyncPrims now takes the render param as a parameter and passes it to the inderivual prim sync functons.  It also now uses the returned dirty bits to clean the prim, rather than always cleaning all bits.
- Added GetRenderParam() to render delegate API.
- Threaded the renderParam through Render Index''s SyncAll()
- Renamed Rprim's Sync() function to _Sync and removed call to GetRepr().  Added new pure virtual Sync().
- Migrated some Rprim schema function to top of the class.
- Sprim and Bprim implementations no longer have to obtain the render index and change tracker from the scene delegate.  Also removed the dirty bit lookup (as now passed as a parameter).  However, these functions are now expected to update the dirty bits with the bits they cleaned.

(Internal change: 1717576)