Process changes before populating selection
This works around a failed verify that would occur
in UsdImagingInstanceAdapter if the parent of a
natively-instanced model were deactivated and
PopulateSelection was called before Render. This
happens to be exactly what usdview does when
deactivating via the right-click context menu.

This is a tactical fix to work around an issue in
a common workflow. We believe UsdImagingDelgate
ultimately needs to provide API or guidance about
what's safe to do when changes are pending.

(Internal change: 1718787)