Fix for memory leak with draw targets.
The workload in this bug report was changing a draw target every frame causing Hydra to recreate the draw target. When running this workload, we were recording a large increase in memory usage.

After examining the objects lifetimes, the Hdx and Glf draw targets lifetimes were being managed correctly. However, when the Glf draw target came to delete the attachments, the attachments were not being deleted because they still had a reference count left. This reference was being held onto by the draw target texture resource in the resource registry. The texture resource was also being apropreatly refcounted down to just 1 reference (the one held by the resource registry itself). The problem was the garbage collector was not being run.

So the memory would have been eventually freed by when we did run the resource registry garbage collector, but the fix is to make sure the garbage collector runs when the draw target needs to be recreated.

Changes:
- HdxDrawTarget::_SetAttachments() now calls SetGarbageCollectionNeeded() on the change tracker if a privious draw target had been set.

(Internal change: 1719866)