Add Commit Resources API
This change is part of the Hydra Render Delegate refactoring.

Currently, HdEngine::Execute() interacts direcctly with the resource registry.  As the resource registry is a RenderDelegate concept (at least, the option whether to use it or not), the calls into it, do not belong in HdEnging.
This change, adds a new Render Delegate API entry point called CommitResources() and places those calls within the stream implementation of it.
Changes:
- Added CommitResources API to Render Delegate
- Added function to get the render delegate from the render index.
- Removed _resourceRegistry from being a member variable of HdEngine.

(Internal change: 1722249)