Don't add spec stack changes due to empty sublayers
Pcp needs to rebuild prim stacks when an empty sublayer
is added to or removed from a layer stack because it
uses indices into the layer stack to represent the
prim stack internally. However, since the sublayer
is empty, the specs in the prim stack don't actually
change, so external clients don't have to be aware
of this.

For these cases, Pcp will now register an 'internal'
spec stack change in the PcpChanges object instead of
a regular one. These internal changes are private to
Pcp, so clients won't be aware of them and won't think
they need to recompute anything.

(Internal change: 1722904)