Windows ArchMakeTmpFile() and ArchMakeTmpSubdir() now match Unix.
Previously the Windows versions wouldn't generate a unique filename
with the same template as on Unix.  This caused the
TfAtomicOfstreamWrapper test to fail because it expected to find
a temporary file with a name matching a particular pattern.

ArchMakeTmpSubdir() didn't even try to make a unique name and
would fail if passed the same arguments twice because it would
try to create the same directory twice.

(Internal change: 1723699)