Made pcp/diagnostic.{cpp,h} private.
A recent change caused diagnostic.h to include debugCodes.h but
debugCodes.h is a private header and cannot be included by a
public header like diagnostic.h.  Furthermore, debugCodes.h
cannot be a public header because there is no support for
exporting the symbols in a TF_DEBUG_CODES.

This change simply makes diagnostic private to pcp.

(Internal change: 1723755)