Add HdxTaskController Interface.
This change is part of the Hydra Render Delegate refactoring.

This change adds the HdxTaskController interface.  As mentioned in a prior checkin that added the HdxRendererPlugin interface, the

Task Controller object, represents a controller and a collection of tasks needed to perform basic operations using the renderer plugin without prior knowledge of the plugin.

Right now, only "bare-bones" functionality is intended to be supported by the task controller.  For example, no lights (and thus no shadows) and no selection or picking support.

Changes:
- Added HdxTaskController interface

(Internal change: 1725203)