Add HdxRendererPlugin interface.
This change is part of the Hydra Render Delegate refactoring.

This changes add's a new interface class to Hdx which represents a Renderer Plugin to the Hydra Ecosystem.  This change is part of an effort
to decouple delegate API's and plugin's, so a plugin can provide multiple delegates.  The Renderer Plugin will be used to factory a render delegate as well as a Task Controller object, which represents a controller and a collection of tasks need to perform basic operations using the renderer plugin without prior knowledge of the plugin.

Changes:
- Added HdxRendererPlugin class.

(Internal change: 1725204)