Add function for computing extents on Boundables
The new UsdGeomBoundable::ComputeExtentFromPlugin
function is a general mechanism for computing extents
given any UsdGeomBoundable prim. To opt-in to this
functionality, schemas must register a function that
will be invoked by UsdGeomBoundable to compute extents
for prims of that type. This is done by using
UsdGeomRegisterComputeExtentFunction.

As the name implies, this function may load plugins
in order to call the registered function.

(Internal change: 1726630)