Relocate hdx prim implementations to HdSt.
The Stream Render delegate factories Hydra Prims for the system.  An upcoming change plans to move HdStreamRenderDelegate to HdSt.  This gives a bit of a library dependency issue as the natural dependency is Hdx depends on HdSt, not the otherway around.

Therefore, this change moves the prims in question (camera, light and draw targets) to HdSt.

As this means touching all the code sites that create these Sprim's.  I upgraded them to the typeId version of InsertSprim and was able to remove the transitional code path.

Changes:
- Moved Camera, Light and Draw Target classes to HdSt.
- Updated names to reflect new library.
- Removed InsertSprim<> function from render index.

(Internal change: 1726929)
(Internal change: 1727001)